# webserwis-net-http/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/crud/": {
            "post": {
                "description": "Receives user data via POST request and returns a success message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "crud"
                ],
                "summary": "Submit user data",
                "parameters": [
                    {
                        "description": "User data payload",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Failed to parse JSON",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Invalid request method",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/crud/{id}": {
            "put": {
                "description": "Receives user data via PUT request and returns a success message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "crud"
                ],
                "summary": "Update user data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User data payload",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Failed to parse JSON",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Invalid request method",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Receives user data via POST request and returns a success message2",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "crud"
                ],
                "summary": "Submit user data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Failed to parse JSON",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Invalid request method",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/file/show": {
            "get": {
                "description": "Endpoint do wyświetlania zawartości pliku w przeglądarce",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "file"
                ],
                "summary": "Display a file",
                "responses": {
                    "200": {
                        "description": "File content displayed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "File not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/hello-world": {
            "get": {
                "description": "Returns a hello world message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "test"
                ],
                "summary": "Hello World Handler",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tickets/history": {
            "get": {
                "description": "Returns ticket history containing searched keywords",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Search keywords in all tickets history",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "List of keywords",
                        "name": "keywords",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tickets/{id}": {
            "get": {
                "description": "Returns ticket",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Get ticket by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tickets/{id}/history": {
            "get": {
                "description": "Returns ticket history",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Get ticket history by ticket id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.RequestBody": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "Adres e-mail użytkownika",
                    "type": "string",
                    "example": "jan.kowalski@example.com"
                },
                "name": {
                    "description": "Imię użytkownika",
                    "type": "string",
                    "example": "Jan Kowalski"
                }
            }
        },
        "model.ResponseBody": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Komunikat do użytkownika",
                    "type": "string",
                    "example": "Data SEND successfully"
                },
                "status": {
                    "description": "Status odpowiedzi",
                    "type": "string",
                    "example": "success"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}